Refactor userReducer to currentUserSlice
