+++ frontend/src/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import userReducer from './reducers/userReducer'
+import currentUserReducer from './features/currentUser/currentUserSlice'
 import feedbacksReducer from './features/feedbacks/feedbacksSlice'
 import commentsReducer from './features/comments/commentsSlice'
 import usersReducer from './features/users/usersSlice'
@@ -7,7 +7,7 @@ import uiReducer from './features/ui/uiSlice'
 
 export const store = configureStore({
   reducer: {
-    currentUser: userReducer,
+    currentUser: currentUserReducer,
     feedbacks: feedbacksReducer,
     comments: commentsReducer,
     users: usersReducer,
